flask app/requirements.txt
flask
numpy
google-cloud-storage